microservices/ocr-service/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.2+cpu
torchvision==0.16.2+cpu
numpy<2
fastapi
uvicorn
easyocr
pillow
pdfplumber
python-docx
python-multipart

